detective/requirements.txt
anthropic
mcp
python-dotenv
google-cloud-bigquery
snowflake-connector-python
cryptography
